speed up

+++ include/Channel/Channel.hpp
@@ -3,7 +3,7 @@
  * @author Sciroccogti (scirocco_gti@yeah.net)
  * @brief
  * @date 2021-03-11 21:03:04
- * @modified: 2021-04-12 01:17:46
+ * @modified: 2021-04-27 13:13:32
  */
 #ifndef CHANNEL_HPP
 #define CHANNEL_HPP
@@ -12,11 +12,11 @@
 
 #include "MatrixMath/MatrixMath.hpp"
 
-Eigen::RowVectorXd AWGN(const Eigen::RowVectorXd& origin, const double snr,
+Eigen::RowVectorXf AWGN(const Eigen::RowVectorXf& origin, const float snr,
                         const int Q, std::default_random_engine engine);
 
-double LLR_AWGN(const double x, const double snr);
-Eigen::MatrixXd LLR_BinAWGN2GF(const Eigen::RowVectorXd& X, const int GF,
-                               const double snr);
+float LLR_AWGN(const float x, const float snr);
+Eigen::MatrixXf LLR_BinAWGN2GF(const Eigen::RowVectorXf& X, const int GF,
+                               const float snr);
 
 #endif
