+++ include/Modem/Modem.hpp
@@ -3,7 +3,7 @@
  * @author Yifan Zhang (scirocco_gti@yeah.net)
  * @brief
  * @date 2020-12-11 13:52:36
- * @modified: 2020-12-11 14:03:59
+ * @modified: 2021-04-27 13:13:38
  */
 
 #ifndef MODEM_HPP
@@ -17,20 +17,20 @@ class Modem {
   private:
     int fc;     // carrier frequency
     int fs;     // sampling frequency
-    double Tb;  // secs of each symbol
+    float Tb;  // secs of each symbol
     int type;   // choose from MODEM_BPSK
     int L;      // number of samples in the time of one symbol
 
   public:
-    Modem(int freqc, int freqs, double Timeb, int t = MODEM_BPSK);
+    Modem(int freqc, int freqs, float Timeb, int t = MODEM_BPSK);
     ~Modem();
-    Eigen::RowVectorXd modulate(const Eigen::RowVectorXi &c);
-    Eigen::RowVectorXd demodulate(const Eigen::RowVectorXd &s);
+    Eigen::RowVectorXf modulate(const Eigen::RowVectorXi &c);
+    Eigen::RowVectorXf demodulate(const Eigen::RowVectorXf &s);
     int getL();
 };
 
 int Compare(const Eigen::MatrixXi &X, const Eigen::MatrixXi &Y);
-// int Compare(const Eigen::MatrixXd &X, const Eigen::MatrixXd &Y);
+// int Compare(const Eigen::MatrixXf &X, const Eigen::MatrixXf &Y);
 int CompareBPSK(const Eigen::RowVectorXi &X, const Eigen::RowVectorXi &Y);
 Eigen::MatrixXi TransBPSK(const Eigen::MatrixXi& X);
 Eigen::MatrixXi RetransBPSK(const Eigen::MatrixXi &X);
