+++ include/UI/pyplot.hpp
@@ -3,7 +3,7 @@
  * @author Yifan Zhang (scirocco_gti@yeah.net)
  * @brief
  * @date 2020-12-13 22:16:26
- * @modified: 2020-12-13 22:16:36
+ * @modified: 2021-04-27 13:10:44
  */
 
 #ifndef PYPLOT_HPP
@@ -12,6 +12,6 @@
 #define PY_SSIZE_T_CLEAN
 #include <Python.h>
 
-int pyplot(double* x, double* y, size_t length, const char* label);
+int pyplot(float* x, float* y, size_t length, const char* label);
 
 #endif
